fix some pr review
