+++ Cargo.toml
@@ -67,7 +67,7 @@ async-stream = "0.3"
 async-trait.workspace = true
 axum = { version = "0.8", features = ["macros"] }
 base64.workspace = true
-cloud-wallet-crypto = { path = "cloud-wallet-crypto" }
+cloud-wallet-crypto = { path = "cloud-wallet-crypto", features = ["jwk"] }
 cloud-wallet-kms = { path = "cloud-wallet-kms", features = ["local-kms"] }
 cloud-wallet-openid4vc = { path = "cloud-wallet-openid4vc" }
 color-eyre.workspace = true
