+++ tests/utils/mod.rs
@@ -1,8 +1,7 @@
-use std::sync::Arc;
+#![allow(dead_code)]
 
-use base64::Engine as _;
-use base64::engine::general_purpose::{STANDARD, URL_SAFE_NO_PAD};
 use cloud_wallet_crypto::ecdsa::{Curve, KeyPair};
+use cloud_wallet_crypto::jwk::Jwk;
 use jsonwebtoken::{Algorithm, EncodingKey, Header, encode};
 use serde_json::json;
 use time::OffsetDateTime;
@@ -10,10 +9,7 @@ use uuid::Uuid;
 
 use cloud_identity_wallet::{
     config::Config,
-    domain::{
-        models::credential::{Credential, CredentialFormat, CredentialStatus},
-        ports::CredentialRepo,
-    },
+    domain::models::credential::{Credential, CredentialFormat, CredentialStatus},
     outbound::MemoryTenantRepo,
     server::Server,
     session::MemorySession,
@@ -25,17 +21,7 @@ use sqlx::{AnyPool, ConnectOptions};
 use time::{OffsetDateTime, UtcDateTime};
 use url::Url;
 
-/// Aborts the server task when dropped, preventing socket/fd leaks between tests.
-#[allow(dead_code)]
-pub struct ServerHandle(tokio::task::JoinHandle<color_eyre::eyre::Result<()>>);
-
-impl Drop for ServerHandle {
-    fn drop(&mut self) {
-        self.0.abort();
-    }
-}
-
-fn make_config() -> Config {
+pub fn make_config() -> Config {
     let mut config = Config::load().unwrap();
     config.server.host = "localhost".to_string();
     config.server.port = 0;
@@ -45,9 +31,7 @@ fn make_config() -> Config {
 
 /// Spawns a test server and returns its base URL.
 ///
-/// The server task runs until the test process exits (fire-and-forget). Use
-/// [`spawn_server_with_repo`] when you need direct repo access or controlled
-/// server teardown.
+/// The server task runs until the test process exits (fire-and-forget).
 pub async fn spawn_server() -> String {
     let config = make_config();
     let session_store = MemorySession::default();
@@ -59,58 +43,17 @@ pub async fn spawn_server() -> String {
     format!("http://{}:{}", config.server.host, port)
 }
 
-/// Spawns a test server and returns the base URL, the shared credential
-/// repository for direct seeding, and a [`ServerHandle`] that aborts the
-/// server task when dropped.
-///
-/// Keep the `ServerHandle` alive for the duration of the test.
-#[allow(dead_code)]
-pub async fn spawn_server_with_repo() -> (String, Arc<dyn CredentialRepo>, ServerHandle) {
-    let config = make_config();
-    let session_store = MemorySession::default();
-    let tenant_repo = MemoryTenantRepo::new();
-    let service = setup::build_service(session_store, tenant_repo, &config).unwrap();
-    let credential_repo = Arc::clone(&service.issuance_engine.credential_repo);
-    let server = Server::new(&config, service).await.unwrap();
-    let port = server.port();
-    let handle = ServerHandle(tokio::spawn(server.run()));
-    (
-        format!("http://{}:{}", config.server.host, port),
-        credential_repo,
-        handle,
-    )
-}
-
 /// Generates a fresh P-256 key pair at runtime.
 ///
 /// Returns an [`EncodingKey`] for signing JWTs and the matching public JWK
 /// value for embedding in the JWT header (so the auth middleware can verify).
-#[allow(dead_code)]
 pub fn create_test_keypair() -> (EncodingKey, serde_json::Value) {
     let key_pair = KeyPair::generate(Curve::P256).expect("P-256 key generation should not fail");
 
-    // jsonwebtoken requires EC private keys as PEM; convert PKCS8 DER → PEM
-    let b64 = STANDARD.encode(key_pair.to_pkcs8_der());
-    let mut pem = String::from("-----BEGIN PRIVATE KEY-----\n");
-    for chunk in b64.as_bytes().chunks(64) {
-        pem.push_str(std::str::from_utf8(chunk).expect("base64 output is always valid ASCII"));
-        pem.push('\n');
-    }
-    pem.push_str("-----END PRIVATE KEY-----\n");
-    let encoding_key = EncodingKey::from_ec_pem(pem.as_bytes())
-        .expect("PEM-encoded EC private key should be valid");
+    let encoding_key = EncodingKey::from_ec_der(key_pair.to_pkcs8_der());
 
-    // Extract uncompressed EC point: 0x04 || x(32 bytes) || y(32 bytes)
-    let mut point = [0u8; 65];
-    key_pair
-        .public_key()
-        .to_sec1_uncompressed(&mut point)
-        .expect("Uncompressed point extraction should not fail");
-
-    let x = URL_SAFE_NO_PAD.encode(&point[1..33]);
-    let y = URL_SAFE_NO_PAD.encode(&point[33..65]);
-
-    let jwk = json!({ "kty": "EC", "crv": "P-256", "x": x, "y": y });
+    let crypto_jwk = Jwk::try_from(&key_pair).expect("P-256 key pair should convert to JWK");
+    let jwk = serde_json::to_value(&crypto_jwk).expect("JWK should serialize to JSON");
 
     (encoding_key, jwk)
 }
